requirements.txt
jsonrpclib-pelix
newspaper3k
numpy==1.19.2
pandas
pika
pymongo
python-dateutil
redis
requests
scipy
sklearn
tensorflow
testresources
watchdog
